requirements.txt
pydantic
discord.py
aiosqlite
jishaku